requirements.txt
streamlit
groq
python-dotenv
pandas
watchdog
pypdf
beautifulsoup4
requests
